refactor(RequiredAuthProvider): move to own file so theres no circular dependency
